codex_telegram_bridge/pyproject.toml
[project]
name = "codex-telegram-bridge"
version = "0.1.0"
description = "Telegram bridge tools for Codex."
readme = "readme.md"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.28.1",
  "markdown-it-py",
  "typer",
]

[project.scripts]
exec-bridge = "codex_telegram_bridge.exec_bridge:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codex_telegram_bridge"]

[tool.hatch.build.targets.sdist]
include = ["src/codex_telegram_bridge", "readme.md"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
